requirements.txt
streamlit>=1.28.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0
streamlit-authenticator>=0.2.3

